# Reduce memory usage when reading payloads that span multiple packets.