--- a/src/MySqlConnector/Protocol/Serialization/IPayloadHandler.cs
+++ b/src/MySqlConnector/Protocol/Serialization/IPayloadHandler.cs
@@ -19,12 +19,15 @@ namespace MySql.Data.Protocol.Serialization
 		/// <summary>
 		/// Reads the next payload.
 		/// </summary>
+		/// <param name="cache">An <see cref="ArraySegmentHolder{Byte}"/> that will cache any buffers allocated during this
+		/// read. (To disable caching, pass <code>new ArraySegmentHolder&lt;byte&gt;</code> so the cache will be garbage-collected
+		/// when this method returns.)</param>
 		/// <param name="protocolErrorBehavior">The <see cref="ProtocolErrorBehavior"/> to use if there is a protocol error.</param>
 		/// <param name="ioBehavior">The <see cref="IOBehavior"/> to use when reading data.</param>
 		/// <returns>An <see cref="ArraySegment{Byte}"/> containing the data that was read. This
 		/// <see cref="ArraySegment{Byte}"/> will be valid to read from until the next time <see cref="ReadPayloadAsync"/> or
 		/// <see cref="WritePayloadAsync"/> is called.</returns>
-		ValueTask<ArraySegment<byte>> ReadPayloadAsync(ProtocolErrorBehavior protocolErrorBehavior, IOBehavior ioBehavior);
+		ValueTask<ArraySegment<byte>> ReadPayloadAsync(ArraySegmentHolder<byte> cache, ProtocolErrorBehavior protocolErrorBehavior, IOBehavior ioBehavior);
 
 		/// <summary>
 		/// Writes a payload.