@bgrainger 2017-04-26
Since the cache is now owned by the `MySqlSession`, it could keep alive `MySqlConnectionStringBuilder.MinimumPoolSize` × *maximum payload size* bytes.

In general, this shouldn't be a problem because idle sessions will eventually get evicted from the pool (to free client and server resources). However, based on the particular workload it could be possible to get into a scenario where many large buffers are being cached longer than necessary. But it's more likely that the caching will improve performance overall for applications that retrieve a lot of large BLOBs from a database.

Finally, it should be noted that Connector/NET has a similar per-connection buffer (`MySqlConnection` holds a `NativeDriver` holds a `MySqlStream` holds a `MySqlPacket` holds a `MemoryStream`) so this is not introducing a new problem specific to this implementation. (It should also be noted that in unscientific local testing I performed, this implementation had 30% fewer bytes in all heaps than Connector/NET when opening a lot of connections, reading a large blob, and keeping all the connections open.)