@bgrainger 2017-04-26
Shipped in [0.19.1](https://www.nuget.org/packages/MySqlConnector/0.19.1).